refactor(code): ADDON-71889: convert jsx files into tsx

+++ ui/src/pages/Input/InputPage.tsx
@@ -1,15 +1,22 @@
-import React, { useState, useEffect, useCallback } from 'react';
+import React, { useState, useEffect, useCallback, ReactElement } from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
 
 import ColumnLayout from '@splunk/react-ui/ColumnLayout';
 import styled from 'styled-components';
 import ToastMessages from '@splunk/react-toast-notifications/ToastMessages';
-import TabBar from '@splunk/react-ui/TabBar';
+import TabBar, { TabBarChangeHandler } from '@splunk/react-ui/TabBar';
 import { _ } from '@splunk/ui-utils/i18n';
+import { z } from 'zod';
+import {
+    InputsPageRegular,
+    InputsPageTableSchema,
+    TableFullServiceSchema,
+    TableLessServiceSchema,
+} from '../../types/globalConfig/pages';
 import { getUnifiedConfigs } from '../../util/util';
 import { TitleComponent, SubTitleComponent } from './InputPageStyle';
-import { TableContextProvider } from '../../context/TableContext';
-import { MODE_CREATE, MODE_CLONE, MODE_EDIT } from '../../constants/modes';
+import { RowDataFields, TableContextProvider } from '../../context/TableContext';
+import { MODE_CREATE, MODE_CLONE, MODE_EDIT, Mode } from '../../constants/modes';
 import { PAGE_INPUT } from '../../constants/pages';
 import { STYLE_PAGE } from '../../constants/dialogStyles';
 import MenuInput, { ROOT_GROUP_NAME } from '../../components/MenuInput/MenuInput';
@@ -36,50 +43,98 @@ const Row = styled(ColumnLayout.Row)`
     }
 `;
 
-function InputPage() {
-    const [entity, setEntity] = useState({ open: false });
+// Define the types based on the Zod schemas
+type InputsPageRegular = z.infer<typeof InputsPageRegular>;
+type InputsPageTableSchema = z.infer<typeof InputsPageTableSchema>;
+
+type InputsPage = {
+    type: 'regular' | 'table';
+} & (InputsPageRegular | InputsPageTableSchema);
+
+type ServiceTableSchema =
+    | z.infer<typeof TableFullServiceSchema>
+    | z.infer<typeof TableLessServiceSchema>;
+
+interface EntityState {
+    open: boolean;
+    isInputPageStyle?: boolean;
+    serviceName: string;
+    mode: Mode;
+    formLabel?: string;
+    groupName?: string;
+    stanzaName?: string;
+}
+
+function isTableFullServiceSchema(
+    service: ServiceTableSchema | undefined
+): service is z.infer<typeof TableFullServiceSchema> {
+    return service !== undefined && 'description' in service;
+}
+
+function InputPage(): ReactElement {
+    const [entity, setEntity] = useState<EntityState>({
+        open: false,
+        serviceName: '',
+        mode: MODE_CREATE,
+    });
     const unifiedConfigs = getUnifiedConfigs();
-    const { services, title, table, description, subDescription } = unifiedConfigs.pages.inputs;
-
-    // check if the tabs feature is enabled or not.
-    const isTabs = !table;
-
-    const [activeTabId, setActiveTabId] = useState(services[0].name);
-    const selectedTab = services.find((x) => x.name === activeTabId);
-
-    const PERMITTED_MODES = [MODE_CLONE, MODE_CREATE, MODE_EDIT];
-    const permittedTabNames = services.map((service) => service.name);
+    const inputsPage = unifiedConfigs.pages.inputs as InputsPage;
+    const [activeTabId, setActiveTabId] = useState<string>('');
 
     const navigate = useNavigate();
     const query = useQuery();
 
     useEffect(() => {
-        // eslint-disable-next-line no-use-before-define
         setServiceEntity();
-        // eslint-disable-next-line no-use-before-define
         setActiveTab();
-    }, [useLocation().search]); // eslint-disable-line react-hooks/exhaustive-deps
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [useLocation().search]);
 
-    const setServiceEntity = () => {
+    let services: ServiceTableSchema[] = [];
+    let title;
+    let table;
+    let description;
+    let subDescription;
+
+    if (inputsPage.type === 'regular') {
+        ({ services, title } = inputsPage);
+        setActiveTabId(services[0].name);
+    } else if (inputsPage.type === 'table') {
+        ({ services, title, table, description, subDescription } = inputsPage);
+        setActiveTabId(services[0].name);
+    }
+
+    const isTabs = !table;
+    const selectedTab = services.find((x) => x.name === activeTabId);
+    const isTableFullSchema = isTableFullServiceSchema(selectedTab);
+
+    const PERMITTED_MODES = [MODE_CLONE, MODE_CREATE, MODE_EDIT];
+    const permittedTabNames = services.map((service) => service.name);
+
+    const setServiceEntity = (): void => {
         const service = services.find((x) => x.name === query.get('service'));
         // Run only when service and action/mode is valid and modal/page is not open
-        if (query && service && PERMITTED_MODES.includes(query.get('action')) && !entity.open) {
+        if (
+            query &&
+            service &&
+            PERMITTED_MODES.includes(query.get('action') || '') &&
+            !entity.open
+        ) {
             // run when mode is not create and previous state info is available
             if (query.get('action') !== MODE_CREATE && entity.stanzaName) {
                 setEntity({
                     ...entity,
                     open: true,
                     isInputPageStyle: true,
-                    serviceName: query.get('service'),
-                    mode: query.get('action'),
+                    serviceName: query.get('service') || '',
+                    mode: (query.get('action') as Mode) || '',
                 });
             } else {
-                // If previous state information is unavailable, create mode will be used by default
                 setEntity({
                     ...entity,
                     open: true,
                     isInputPageStyle: true,
-                    serviceName: query.get('service'),
+                    serviceName: query.get('service') || '',
                     formLabel: `Create ${service?.title}`,
                     mode: MODE_CREATE,
                 });
@@ -89,22 +144,28 @@ function InputPage() {
             entity.open &&
             entity.isInputPageStyle
         ) {
-            // Close page when any of the required query params are not provided
             setEntity({ ...entity, open: false });
         }
     };
 
-    const setActiveTab = () => {
-        if (query && permittedTabNames.includes(query.get('service'))) {
-            setActiveTabId(query.get('service'));
+    const setActiveTab = (): void => {
+        if (query && permittedTabNames.includes(query.get('service') || '')) {
+            setActiveTabId(query.get('service') || '');
         }
     };
 
-    // handle modal/page open request on create/add entity button
-    const handleRequestOpen = ({ serviceName, groupName, input }) => {
+    const handleRequestOpen = ({
+        serviceName,
+        groupName,
+        input,
+    }: {
+        serviceName: string;
+        groupName?: string;
+        input?: string;
+    }): void => {
         const service = services.find((x) => x.name === serviceName);
-        const serviceTitle = service.title;
-        const isInputPageStyle = service.style === STYLE_PAGE;
+        const serviceTitle = service?.title;
+        const isInputPageStyle = service?.style === STYLE_PAGE;
 
         setEntity({
             ...entity,
@@ -116,7 +177,6 @@ function InputPage() {
             groupName,
         });
         if (isInputPageStyle) {
-            // set query and push to navigate
             query.set('service', serviceName);
             query.set('action', MODE_CREATE);
             const selectedGroup = groupName && groupName !== ROOT_GROUP_NAME ? groupName : null;
@@ -130,13 +190,11 @@ function InputPage() {
         }
     };
 
-    // handle close/cancel/back request in add/create modal component
-    const handleModalDialogClose = () => {
+    const handleModalDialogClose = (): void => {
         setEntity({ ...entity, open: false });
     };
 
-    // generate modal style dialog
-    const generateModalDialog = () => (
+    const generateModalDialog = (): ReactElement => (
         <EntityModal
             page={PAGE_INPUT}
             open={entity.open}
@@ -148,8 +206,7 @@ function InputPage() {
         />
     );
 
-    // handle clone/edit request per row from table for page style dialog
-    const handleOpenPageStyleDialog = (row, mode) => {
+    const handleOpenPageStyleDialog = (row: RowDataFields, mode: Mode): void => {
         const label = services.find((x) => x.name === row.serviceName)?.title;
         setEntity({
             ...entity,
@@ -158,16 +215,14 @@ function InputPage() {
             serviceName: row.serviceName,
             stanzaName: row.name,
             formLabel: mode === MODE_CLONE ? `Clone ${label}` : `Update ${label}`,
-            mode,
+            mode: mode as Mode,
         });
-        // set query and push to history
         query.set('service', row.serviceName);
         query.set('action', mode);
         navigate({ search: query.toString() });
     };
 
-    // handle close request for page style dialog
-    const handlePageDialogClose = () => {
+    const handlePageDialogClose = (): void => {
         setEntity({ ...entity, open: false });
         if (!isTabs) {
             query.delete('service');
@@ -176,10 +231,8 @@ function InputPage() {
         navigate({ search: query.toString() });
     };
 
-    // generate page style dialog
-    const generatePageDialog = () => (
+    const generatePageDialog = (): ReactElement => (
         <EntityPage
-            open={entity.open}
             handleRequestClose={handlePageDialogClose}
             serviceName={entity.serviceName}
             stanzaName={entity.stanzaName}
@@ -190,20 +243,21 @@ function InputPage() {
         />
     );
 
-    const onTabChange = useCallback(
+    const onTabChange = useCallback<TabBarChangeHandler>(
         (e, { selectedTabId }) => {
-            setActiveTabId(selectedTabId);
-            query.delete('action');
-            query.set('service', selectedTabId);
-            navigate({ search: query.toString() });
-            // eslint-disable-next-line react-hooks/exhaustive-deps
+            if (selectedTabId) {
+                setActiveTabId(selectedTabId);
+                query.delete('action');
+                query.set('service', selectedTabId);
+                navigate({ search: query.toString() });
+            }
         },
         [activeTabId] // eslint-disable-line react-hooks/exhaustive-deps
     );
 
     return (
         <ErrorBoundary>
-            <TableContextProvider value={null}>
+            <TableContextProvider>
                 {entity.isInputPageStyle && entity.open ? generatePageDialog() : null}{' '}
                 <div
                     style={
@@ -214,14 +268,22 @@ function InputPage() {
                 >
                     <ColumnLayout gutter={8}>
                         <Row>
-                            <ColumnLayout.Column className={isTabs && 'title_menu_column'} span={9}>
+                            <ColumnLayout.Column
+                                className={isTabs ? 'title_menu_column' : ''}
+                                span={9}
+                            >
                                 <TitleComponent>
-                                    {isTabs ? _(selectedTab.title) : _(title || '')}
+                                    {isTabs ? _(selectedTab?.title || '') : _(title || '')}
                                 </TitleComponent>
-                                <SubTitleComponent className={isTabs && 'page_subtitle'}>
-                                    {isTabs ? _(selectedTab.description) : _(description || '')}
+                                <SubTitleComponent className={isTabs ? 'pageSubtitle' : undefined}>
+                                    {isTabs
+                                        ? _(isTableFullSchema ? selectedTab?.description : '')
+                                        : _(description || '')}
                                 </SubTitleComponent>
-                                <SubDescription {...subDescription} />
+                                <SubDescription
+                                    text={subDescription?.text || ''}
+                                    links={subDescription?.links}
+                                />
                             </ColumnLayout.Column>
                             <ColumnLayout.Column
                                 className={isTabs ? 'title_menu_column' : 'dropdown'}
@@ -259,6 +321,7 @@ function InputPage() {
                                             handleRequestOpen({ serviceName: service.name })
                                         }
                                         handleOpenPageStyleDialog={handleOpenPageStyleDialog}
+                                        displayActionBtnAllRows
                                     />
                                 </div>
                             ))}
@@ -268,6 +331,7 @@ function InputPage() {
                             page={PAGE_INPUT}
                             handleOpenPageStyleDialog={handleOpenPageStyleDialog}
                             displayActionBtnAllRows
+                            serviceName=""
                         />
                     )}
